Split up editableTree Files
