+++ packages/dds/tree/src/feature-libraries/editable-tree/index.ts
@@ -8,8 +8,6 @@ export {
 	EditableTree,
 	EditableField,
 	EditableTreeOrPrimitive,
-	isEditableField,
-	isUnwrappedNode,
 	proxyTargetSymbol,
 	UnwrappedEditableTree,
 	UnwrappedEditableField,
@@ -19,7 +17,10 @@ export {
 	parentField,
 	EditableTreeEvents,
 	on,
-} from "./editableTree";
+} from "./editableTreeTypes";
+
+export { isEditableField } from "./editableField";
+export { isUnwrappedNode } from "./editableTree";
 
 export { EditableTreeContext, getEditableTreeContext } from "./editableTreeContext";
