+++ packages/dds/tree/src/test/feature-libraries/editable-tree/mockData.ts
@@ -14,6 +14,13 @@ import {
 	ContextuallyTypedNodeDataObject,
 	jsonableTreeFromCursor,
 	cursorFromContextualData,
+	EditableTreeContext,
+	DefaultEditBuilder,
+	ContextuallyTypedNodeData,
+	buildForest,
+	cursorsFromContextualData,
+	defaultSchemaPolicy,
+	getEditableTreeContext,
 } from "../../../feature-libraries";
 import {
 	ValueSchema,
@@ -26,6 +33,11 @@ import {
 	GlobalFieldKeySymbol,
 	FieldSchema,
 	SchemaData,
+	IEditableForest,
+	SchemaDataAndPolicy,
+	InMemoryStoredSchemaRepository,
+	lookupGlobalFieldSchema,
+	initializeForest,
 } from "../../../core";
 import { brand, Brand } from "../../../util";
 
@@ -289,4 +301,40 @@ export function buildTestSchema(rootField: FieldSchema = rootPersonSchema): Sche
 		new Map([...fullSchemaData.globalFieldSchema.entries(), [rootFieldKey, rootField]]),
 		...treeSchema,
 	);
+}
+
+export function getReadonlyEditableTreeContext(forest: IEditableForest): EditableTreeContext {
+	// This will error if someone tries to call mutation methods on it
+	const dummyEditor = {} as unknown as DefaultEditBuilder;
+	return getEditableTreeContext(forest, dummyEditor);
+}
+
+export function setupForest(
+	schema: SchemaData,
+	data: ContextuallyTypedNodeData | undefined,
+): IEditableForest {
+	const schemaRepo = new InMemoryStoredSchemaRepository(defaultSchemaPolicy, schema);
+	const forest = buildForest(schemaRepo);
+	const root = cursorsFromContextualData(
+		schemaRepo,
+		lookupGlobalFieldSchema(schemaRepo, rootFieldKey),
+		data,
+	);
+	initializeForest(forest, root);
+	return forest;
+}
+
+export function buildTestTree(
+	data: ContextuallyTypedNodeData | undefined,
+	rootField: FieldSchema = rootPersonSchema,
+): EditableTreeContext {
+	const schema: SchemaData = buildTestSchema(rootField);
+	const forest = setupForest(schema, data);
+	const context = getReadonlyEditableTreeContext(forest);
+	return context;
+}
+
+export function buildTestPerson(): readonly [SchemaDataAndPolicy, Person] {
+	const context = buildTestTree(personData);
+	return [context.schema, context.unwrappedRoot as Person];
 }
